+++ client/src/game/api/events/client.ts
@@ -1,4 +1,4 @@
-import { ServerClient } from "../../comm/types/settings";
+import { ServerClient, userOptionsToClient } from "../../comm/types/settings";
 import { EventBus } from "../../event-bus";
 import { layerManager } from "../../layers/manager";
 import { gameStore } from "../../store";
@@ -9,14 +9,27 @@ socket.on("Client.Options.Set", (options: ServerClient) => {
     const path = window.location.pathname.split("/");
     gameStore.setDM(options.name === decodeURIComponent(path[path.length - 2]));
 
-    gameStore.setGridSize({ gridSize: options.grid_size, sync: false });
-    gameStore.setGridColour({ colour: options.grid_colour, sync: false });
-    gameStore.setFOWColour({ colour: options.fow_colour, sync: false });
-    gameStore.setRulerColour({ colour: options.ruler_colour, sync: false });
-    gameStore.setInvertAlt({ invertAlt: options.invert_alt, sync: false });
-    gameStore.setPanX(options.pan_x);
-    gameStore.setPanY(options.pan_y);
-    gameStore.setZoomDisplay(options.zoom_factor);
+    gameStore.setDefaultClientOptions(userOptionsToClient(options.default_user_options));
+
+    if (options.room_user_options?.grid_size)
+        gameStore.setGridSize({ gridSize: options.room_user_options.grid_size, sync: false });
+    else gameStore.setGridSize({ gridSize: options.default_user_options.grid_size, sync: false });
+    if (options.room_user_options?.grid_colour)
+        gameStore.setGridColour({ colour: options.room_user_options.grid_colour, sync: false });
+    else gameStore.setGridColour({ colour: options.default_user_options.grid_colour, sync: false });
+    if (options.room_user_options?.fow_colour)
+        gameStore.setFOWColour({ colour: options.room_user_options.fow_colour, sync: false });
+    else gameStore.setFOWColour({ colour: options.default_user_options.fow_colour, sync: false });
+    if (options.room_user_options?.ruler_colour)
+        gameStore.setRulerColour({ colour: options.room_user_options.ruler_colour, sync: false });
+    else gameStore.setRulerColour({ colour: options.default_user_options.ruler_colour, sync: false });
+    if (options.room_user_options?.invert_alt)
+        gameStore.setInvertAlt({ invertAlt: options.room_user_options.invert_alt, sync: false });
+    else gameStore.setInvertAlt({ invertAlt: options.default_user_options.invert_alt, sync: false });
+
+    gameStore.setPanX(options.location_user_options.pan_x);
+    gameStore.setPanY(options.location_user_options.pan_y);
+    gameStore.setZoomDisplay(options.location_user_options.zoom_factor);
 
     EventBus.$once("Board.Floor.Set", () => {
         if (options.active_layer) layerManager.selectLayer(options.active_layer, false);