+++ server/api/socket/location.py
@@ -81,9 +81,13 @@ async def load_location(sid: str, location: Location, *, complete=False):
     # 1. Load client options
 
     client_options = pr.player.as_dict()
-    client_options.update(
-        **LocationUserOption.get(user=pr.player, location=location).as_dict()
-    )
+    client_options["location_user_options"] = LocationUserOption.get(
+        user=pr.player, location=location
+    ).as_dict()
+    client_options["default_user_options"] = pr.player.default_options.as_dict()
+
+    if pr.user_options:
+        client_options["room_user_options"] = pr.user_options.as_dict()
 
     await sio.emit("Client.Options.Set", client_options, room=sid, namespace=GAME_NS)