+++ server/models/user.py
@@ -1,22 +1,45 @@
 import bcrypt
-from peewee import fn, BooleanField, IntegerField, TextField
+from peewee import ForeignKeyField, fn, BooleanField, IntegerField, TextField
 from playhouse.shortcuts import model_to_dict
 
 from .base import BaseModel
 
 
-__all__ = ["User"]
+__all__ = ["User", "UserOptions"]
+
+
+class UserOptions(BaseModel):
+    fow_colour = TextField(default="#000", null=True)
+    grid_colour = TextField(default="#000", null=True)
+    ruler_colour = TextField(default="#F00", null=True)
+    invert_alt = BooleanField(default=False, null=True)
+    grid_size = IntegerField(default=50, null=True)
+
+    @classmethod
+    def create_empty(cls):
+        return UserOptions.create(
+            fow_colour=None,
+            grid_colour=None,
+            ruler_colour=None,
+            invert_alt=None,
+            grid_size=None,
+        )
+
+    def as_dict(self):
+        return {
+            k: v
+            for k, v in model_to_dict(
+                self, backrefs=None, recurse=None, exclude=[UserOptions.id]
+            ).items()
+            if v is not None
+        }
 
 
 class User(BaseModel):
     name = TextField()
     email = TextField(null=True)
     password_hash = TextField()
-    fow_colour = TextField(default="#000")
-    grid_colour = TextField(default="#000")
-    ruler_colour = TextField(default="#F00")
-    invert_alt = BooleanField(default=False)
-    grid_size = IntegerField(default=50)
+    default_options = ForeignKeyField(UserOptions, on_delete="CASCADE")
 
     def __repr__(self):
         return f"<User {self.name}>"
@@ -32,7 +55,11 @@ class User(BaseModel):
         return bcrypt.checkpw(pw.encode("utf8"), expected_hash)
 
     def as_dict(self):
-        return model_to_dict(self, recurse=False, exclude=[User.id, User.password_hash])
+        return model_to_dict(
+            self,
+            recurse=False,
+            exclude=[User.id, User.password_hash, User.default_options],
+        )
 
     @classmethod
     def by_name(cls, name):