+++ CHANGELOG.md
@@ -8,6 +8,12 @@ These usually have no immediately visible impact on regular users
 
 ## Unreleased
 
+### Changed
+
+-   Client settings
+    -   Now open in a modal just like the DM settings
+    -   Now can be campaign-specific (with reset/make default options)
+
 ### Fixed
 
 -   It's no longer possible to create a floor with a name that is already in use